feat: Reconfig and adding User entity to backend.
[Domain]
Added User entity, DTOs, mapper.
Renamed field names in various entities.
Changed all DTOs to readonly.
Changed mappers according to changes.
Removed SortBy from Product functions.

[Application]
Added and edited custom exceptions.
Added IUserRepository and UserService.
Removed unused methods from repository interfaces.
Added XML documentation comments to methods.
Refactored services accordingly to the Domain project changes.
Services now take DTOs as method parameters instead of primitives.

[Infrastructure]
Refactored repositories accordingly to other project's changes.
Added XML documentation comments to methods.
Added UserRepository.

[WebAPI]
Refactored controllers accordingly to other project's changes.
Renamed various queries and reordered argument positions for aethestic.
Added UserController.
Added Console logging (Information level).
Added Cors default policy to accept all.

[Docker]
Changed to use external volume instead of mapping path.

+++ src/Domain/DTOs/Invoices/InvoiceShort.cs
@@ -5,6 +5,6 @@ public readonly record struct InvoiceShort(
 	string? ClientName,
 	DateTime DateCreated,
 	uint GrandTotal,
-	ushort TotalProduct,
+	uint TotalProduct,
 	string FirstProductName,
 	uint FirstProductQuantity);
